Menambahkan Autentikasi Role User

+++ dashboard/proses.php
@@ -1,5 +1,5 @@
 <?php
-session_start();
+include '../utilities/validate.php';
 include '../config/koneksi.php';
 
 
@@ -13,6 +13,12 @@ if(isset($_POST['submit_jenis_obat'])){
     $kategori = $_POST['cat'];
     $deskripsi = $_POST['desc'];
     $efekSamping = $_POST['efek'];
+
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "INSERT INTO `category` (`kategori`, `deskripsi`, `efekSamping`) VALUES ('$kategori', '$deskripsi', '$efekSamping')");
 
     // if(!preg_match("/^[a-zA-Z-']*$/",$kategori) || $kategori == "" || $deskripsi == "" || $efekSamping = ""){
@@ -29,6 +35,11 @@ elseif(isset($_POST['edit_jenis_obat'])){
     $kategori = $_POST['cat'];
     $deskripsi = $_POST['desc'];
     $efekSamping = $_POST['efek'];
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "UPDATE `category` SET kategori='$kategori', deskripsi='$deskripsi', efekSamping='$efekSamping' WHERE id='$id'") or die(mysqli_error($koneksi));
     if($query){
         $_SESSION['message'] = "<script>Swal.fire({title: 'Berhasil!',text: 'Data Berhasil Di Edit!',icon: 'success',confirmButtonText: 'OK'})</script>";
@@ -38,6 +49,11 @@ elseif(isset($_POST['edit_jenis_obat'])){
 elseif(isset($_POST['hapus_category'])){
     $id = $_POST['id'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "DELETE FROM category WHERE id='$id'") or die(mysqli_error($koneksi));
 
     if($query){
@@ -49,6 +65,11 @@ elseif(isset($_POST['hapus_category'])){
 elseif(isset($_POST['submit_unit_obat'])){
     $unit = $_POST['unit'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "INSERT INTO units (unit) VALUES ('$unit')") or die(mysqli_error($koneksi));
     if($query){
         $_SESSION['message'] = "<script>Swal.fire({title: 'Berhasil!',text: 'Data Berhasil Di Tambahkan!',icon: 'success',confirmButtonText: 'OK'})</script>";
@@ -59,6 +80,11 @@ elseif(isset($_POST['edit_unit_obat'])){
     $id = $_POST['id'];
     $unit = $_POST['unit'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "UPDATE `units` SET unit='$unit' WHERE id='$id'") or die(mysqli_error($koneksi));
     if($query){
         $_SESSION['message'] = "<script>Swal.fire({title: 'Berhasil!',text: 'Data Berhasil Di Ubah!',icon: 'success',confirmButtonText: 'OK'})</script>";
@@ -71,6 +97,11 @@ elseif(isset($_POST['edit_unit_obat'])){
 elseif(isset($_POST['hapus_unit'])){
     $id = $_POST['id'];
     
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "DELETE FROM units WHERE id='$id'") or die(mysqli_error($koneksi));
 
     if($query){
@@ -82,18 +113,24 @@ elseif(isset($_POST['hapus_unit'])){
         echo "Gagal";
     }
 }
+
 elseif(isset($_POST['submit_new_user'])){
     $nama = $_POST['npegawai'];
     $username = $_POST['username'];
     $password = md5($_POST['pass']);
     $level = $_POST['level'];
 
+    if($_SESSION['level'] != 1337){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     if($level < 1 || $level > 2){
         $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Silahkan Pilih Level User!',icon: 'error',confirmButtonText: 'OK'})</script>";
         header("Location: index.php");
     }
     else{
-        $query = mysqli_query($koneksi, "INSERT INTO `users` (`username`, `password`, `level`) VALUES ('$username', '$password', '$level')") or die(mysqli_error($koneksi));
+        $query = mysqli_query($koneksi, "INSERT INTO `users` (`nama_pegawai`, `username`, `password`, `level`) VALUES ('$nama', '$username', '$password', '$level')") or die(mysqli_error($koneksi));
         $_SESSION['message'] = "<script>Swal.fire({title: 'Berhasil!',text: 'Berhasil Menambahkan Pegawai',icon: 'success',confirmButtonText: 'OK'})</script>";
         header("Location: index.php");
     }
@@ -101,6 +138,11 @@ elseif(isset($_POST['submit_new_user'])){
 elseif(isset($_POST['submit_pemasok'])){
     $nama = $_POST['nama'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "INSERT INTO pemasok (nama) VALUES ('$nama')");
 
     if($query){
@@ -123,6 +165,12 @@ elseif(isset($_POST['submit_obat'])){
     $hargajual = $_POST['hargajual'];
     $pemasok = $_POST['pemasok'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
+
     $query = mysqli_query($koneksi, "INSERT INTO obat (nama, penyimpanan, Stock, Unit, kategori, t_kadaluarsa, deskripsi, h_beli, h_jual, pemasok) VALUES ('$nama', '$penyimpanan', $stock, '$unit', '$kategori', '$kadaluarsa', '$deskripsi', $hargabeli, $hargajual, '$pemasok')");
 
     if($query){
@@ -143,6 +191,11 @@ elseif(isset($_POST['edit_obat'])){
     $hargajual = $_POST['hargajual'];
     $pemasok = $_POST['pemasok'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "UPDATE `obat` SET nama='$nama', penyimpanan='$penyimpanan', Stock='$stock', Unit='$unit', kategori='$kategori', t_kadaluarsa='$kadaluarsa', deskripsi='$deskripsi', h_beli='$hargabeli', h_jual='$hargajual', pemasok='$pemasok' WHERE id='$id'") or die(mysqli_error($koneksi));;
 
     if($query){
@@ -157,6 +210,11 @@ elseif(isset($_POST['edit_obat'])){
 elseif(isset($_POST['hapus_obat'])){
     $id = $_POST['id'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "DELETE FROM obat WHERE id='$id'");
 
     if($query){
@@ -171,6 +229,11 @@ elseif(isset($_POST['edit_pemasok'])){
     $id = $_POST['id'];
     $nama = $_POST['pemasok'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "UPDATE pemasok SET nama='$nama' WHERE id='$id'");
 
     if($query){
@@ -185,6 +248,11 @@ elseif(isset($_POST['edit_pemasok'])){
 elseif(isset($_POST['hapus_pemasok'])){
     $id = $_POST['id'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "DELETE FROM pemasok WHERE id='$id'");
 
     if($query){
@@ -201,6 +269,11 @@ elseif(isset($_POST['edit_pegawai'])){
     $username = $_POST['username'];
     $level = $_POST['level'];
 
+    if($_SESSION['level'] != 1337){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "UPDATE users SET nama_pegawai='$namapegawai', username='$username', level='$level' WHERE id='$id'");
 
     if($query){
@@ -214,6 +287,11 @@ elseif(isset($_POST['edit_pegawai'])){
 elseif(isset($_POST['hapus_pegawai'])){
     $id = $_POST['id'];
 
+    if($_SESSION['level'] != 1337){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = mysqli_query($koneksi, "DELETE FROM users WHERE id='$id'");
 
     if($query){
@@ -234,6 +312,11 @@ elseif(isset($_POST['submit_penjualan'])){
     $total = $_POST['total'];
     $newStock = $stock - $banyakydbeli;
 
+    if($_SESSION['level'] == 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = "INSERT INTO penjualan (nama, obat, unit, harga, beli, total) VALUES ('$namapembeli', '$obat', '$unit', $harga, $banyakydbeli, $total);
     UPDATE obat SET Stock='$newStock' WHERE nama='$obat'";
 
@@ -248,6 +331,11 @@ elseif(isset($_POST['submit_penjualan'])){
 elseif(isset($_POST['hapus_penjualan'])){
     $id = $_POST['id'];
 
+    if($_SESSION['level'] == 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = "DELETE FROM penjualan WHERE id='$id'";
 
     if($koneksi->query($query)){
@@ -269,6 +357,11 @@ elseif(isset($_POST['submit_pembelian'])){
     $created = date('Y-m-d');
     $newStock = $stock + $banyak_stock;
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = "INSERT INTO pembelian (pemasok, t_transaksi, obat, harga, banyak_stock, total, created) VALUES ('$pemasok', '$t_transaksi', '$obat', $harga, $banyak_stock, $total, '$created');
     UPDATE obat SET Stock='$newStock' WHERE nama='$obat' AND pemasok='$pemasok'";
 
@@ -283,6 +376,11 @@ elseif(isset($_POST['submit_pembelian'])){
 elseif(isset($_POST['hapus_pembelian'])){
     $id = $_POST['id'];
 
+    if($_SESSION['level'] < 2){
+        $_SESSION['message'] = "<script>Swal.fire({title: 'Error!',text: 'Kamu Tidak Punya Akses!',icon: 'error',confirmButtonText: 'OK'})</script>";
+        header("Location: index.php");
+        die();
+    }
     $query = "DELETE FROM pembelian WHERE id='$id'";
 
     if($koneksi->query($query)){
